add item view for Create, now using a simple Label method, since Item no longer has an explicitly defined Category property.
Remember that we use the HTML helper method DropDownList() to create an HTML dropdown menu consisting of a select element with nested option elements. The user will see a Category's name for each option while the actual value of the option will be the CategoryId. This is because of the SelectList we created in the controller method.